% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALDataset}
\alias{GDALDataset}
\title{get a pointer to the abstract GDALDataset (not much use, see gdalptr::GDALOpen)}
\usage{
GDALDataset()
}
\description{
get a pointer to the abstract GDALDataset (not much use, see gdalptr::GDALOpen)
}
