% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALGetInfo}
\alias{GDALGetInfo}
\title{Get the number of columns of  data set opened with gdalptr::GDALOpen()}
\usage{
GDALGetInfo(x)
}
\arguments{
\item{x}{external pointer returned by gdalptr::GDALOpen()}
}
\description{
Get the number of columns of  data set opened with gdalptr::GDALOpen()
}
