% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALGetRasterSize}
\alias{GDALGetRasterSize}
\title{Get the number of columns of  data set opened with gdalptr::GDALOpen()}
\usage{
GDALGetRasterSize(x)
}
\arguments{
\item{x}{external pointer returned by gdalptr::GDALOpen()}
}
\value{
integer vector ncol,nrow
}
\description{
Get the number of columns of  data set opened with gdalptr::GDALOpen()
}
\examples{
dsn <- system.file("extdata/volcano_gcp.tif", package = "gdalptr", mustWork  = TRUE)
ds <- GDALOpen(dsn)
GDALGetRasterSize(ds)
}
