% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALOpen}
\alias{GDALOpen}
\title{Title}
\usage{
GDALOpen(dsn = NULL)
}
\arguments{
\item{a}{}
}
\description{
Title
}
\examples{
dsn <- system.file("extdata/volcano_gcp.tif", package = "gdalptr", mustWork  = TRUE)
ds <- GDALOpen(dsn)
GDALGetInfo(ds)
}
