% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr-package.R
\docType{package}
\name{gdalptr-package}
\alias{gdalptr}
\alias{gdalptr-package}
\title{gdalptr: Pointer Fun with Mikey}
\description{
Just messing around learning from paleolimbot/radbc.
}
\author{
\strong{Maintainer}: Michael Sumner \email{first.last@example.com}

}
\keyword{internal}
