% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALVersion}
\alias{GDALVersion}
\title{Get the GDAL version string}
\usage{
GDALVersion()
}
\value{
character vector describing GDAL version
}
\description{
Get the GDAL version string
}
\examples{
GDALVersion()
}
