% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALVSIReadDirRecursive}
\alias{GDALVSIReadDirRecursive}
\title{List VSI recursively}
\usage{
GDALVSIReadDirRecursive(vsiname)
}
\arguments{
\item{vsiname}{character string to GDAL source}
}
\value{
character vector of source contents
}
\description{
I.e. get the file list from a zip file, so we can construct a dsn.
}
\examples{
url <- file.path("/vsitar//vsicurl", 
       "https://cran.r-project.org/src/contrib/Archive/MASS/MASS_7.3-0.tar.gz")
GDALVSIReadDirRecursive(url)
}
