% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xptr_GDALGetRasterSize}
\alias{xptr_GDALGetRasterSize}
\title{Query the raster size a gdal dataset}
\usage{
xptr_GDALGetRasterSize(xp)
}
\arguments{
\item{xp}{external pointer returned by \code{\link[=xptr_GDALOpen]{xptr_GDALOpen()}}}
}
\value{
ncol, nrow of underlying data set
}
\description{
Query the raster size a gdal dataset
}
