% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xptr_GDALOpen}
\alias{xptr_GDALOpen}
\title{Open a gdal dataset}
\usage{
xptr_GDALOpen(dsn, sds = as.integer(c(0)))
}
\arguments{
\item{dsn}{data source name (file path, url,etc)}

\item{sds}{subdataset to choose (1-based, or defaults to 1 if not specified or if there are no sds)}
}
\value{
external pointer
}
\description{
Open a gdal dataset
}
