% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalptr.R
\name{GDALVSIReadDirRecursive}
\alias{GDALVSIReadDirRecursive}
\title{List VSI recursively}
\usage{
GDALVSIReadDirRecursive(vsiname)
}
\arguments{
\item{vsiname}{character string to GDAL source}
}
\value{
character vector of source contents
}
\description{
I.e. get the file list from a zip file, so we can construct a dsn.
}
\examples{
url <- file.path("/vsitar//vsicurl", 
       "https://cran.r-project.org/src/contrib/Archive/pixmap/pixmap_0.4-11.tar.gz")
if (interactive()) {
  files <- try(GDALVSIReadDirRecursive(url))
  if (!inherits(files, "try-error")) {
  dsn <- file.path(url, grep("logo.ppm", files, value = TRUE))
   ppm <- GDALOpen(dsn)
   GDALGetRasterSize(ppm) 
   
   }
  }
}
